feat(monitor): relocate report helpers → session/monitor/report
Moves display_op_trace_report + write_op_trace_json. Old path retained
as re-export shim. Existing test file moved with git mv (history preserved).
